+++ src/main/java/com/example/employee_management_system/model/Employee.java
@@ -4,6 +4,8 @@ import jakarta.persistence.*;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
 
+import java.math.BigDecimal;
+
 @EqualsAndHashCode(callSuper = true)
 @Entity
 @Data
@@ -15,7 +17,7 @@ public class Employee extends Person {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long employeeId;
-    private Double salary;
+    private BigDecimal salary;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "employee_function_id")
